backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncmy
python-dotenv
passlib[bcrypt]
PyJWT